+++ src/components/azure-pg/index.test.ts
@@ -4,8 +4,6 @@ import { createNodeCJSEnvironment } from "@kosko/env";
 import { promises } from "fs";
 import { directory } from "tempy";
 
-import { create } from "./index";
-
 const gitlabEnv = {
   CI_COMMIT_SHORT_SHA: "abcdefg",
   CI_ENVIRONMENT_NAME: "fabrique-dev",
@@ -17,10 +15,8 @@ const gitlabEnv = {
   KUBE_NAMESPACE: "sample-42-my-test",
 };
 
-jest.mock("@socialgouv/kosko-charts/environments/gitlab", () => ({
-  // eslint-disable-next-line @typescript-eslint/naming-convention
-  __esModule: true,
-  default: () => ({
+const gitlabMock = {
+  metadata: {
     annotations: {
       "app.gitlab.com/app": "socialgouv-sample",
       "app.gitlab.com/env": "my-test",
@@ -31,19 +27,32 @@ jest.mock("@socialgouv/kosko-charts/environments/gitlab", () => ({
       team: "sample",
     },
     namespace: { name: "sample-42-my-test" },
-  }),
-}));
+  },
+  projectName: "sample",
+  shortSha: "abcdefg",
+};
+
+const gitlabModuleMock = {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  __esModule: true,
+  default: () => gitlabMock,
+};
 
 beforeEach(() => {
   jest.resetModules();
 });
 
-test("should throw because of a missing envs", () => {
+test("should throw because of a missing envs", async () => {
   const env = createNodeCJSEnvironment({ cwd: "/tmp" });
+  const { create } = await import("./index");
   expect(() => create({ env })).toThrowErrorMatchingSnapshot();
 });
 
 test("should return create an job", async () => {
+  jest.doMock(
+    "@socialgouv/kosko-charts/environments/gitlab",
+    () => gitlabModuleMock
+  );
   Object.assign(process.env, gitlabEnv);
   const cwd = directory();
   const env = createNodeCJSEnvironment({ cwd });
@@ -53,10 +62,15 @@ test("should return create an job", async () => {
     `${cwd}/environments/dev/pg.sealed-secret.yaml`,
     "---\napiVersion: v1\nkind: ConfigMap"
   );
+  const { create } = await import("./index");
   expect(create({ env })).toMatchSnapshot();
 });
 
 test("should use custom pgHost", async () => {
+  jest.doMock(
+    "@socialgouv/kosko-charts/environments/gitlab",
+    () => gitlabModuleMock
+  );
   process.env.CI_PROJECT_NAME = "sample-next-app";
   const env = createNodeCJSEnvironment({ cwd: "/tmp/xxx" });
   env.env = "dev";
@@ -65,6 +79,7 @@ test("should use custom pgHost", async () => {
     `/tmp/xxx/environments/dev/pg.sealed-secret.yaml`,
     "---\napiVersion: v1\nkind: ConfigMap"
   );
+  const { create } = await import("./index");
   expect(
     create({ config: { pgHost: "pouetpouet.com" }, env })
   ).toMatchSnapshot();