+++ src/components/namespace/index.ts
@@ -1,4 +1,4 @@
-import gitlab from "@socialgouv/kosko-charts/environments/gitlab";
+import environments from "@socialgouv/kosko-charts/environments";
 import type { NamespaceComponentEnvironment } from "@socialgouv/kosko-charts/types";
 import { merge } from "@socialgouv/kosko-charts/utils/@kosko/env/merge";
 import { Namespace as K8SNamespace } from "kubernetes-models/v1/Namespace";
@@ -6,10 +6,10 @@ import { Namespace as K8SNamespace } from "kubernetes-models/v1/Namespace";
 export const createNamespace = (
   config?: NamespaceComponentEnvironment
 ): K8SNamespace => {
-  const gitlabEnv = gitlab(process.env);
-  const owner = gitlabEnv.labels?.owner;
+  const gitlabEnv = environments(process.env);
+  const owner = gitlabEnv.metadata.labels?.owner;
 
-  const envParams = merge(gitlab(process.env), config ?? {});
+  const envParams = merge(gitlabEnv.metadata, config ?? {});
 
   const namespace = new K8SNamespace({
     metadata: {