+++ src/utils/createIngress.test.ts
@@ -1,28 +1,23 @@
-import createIngress from "./createIngress";
-
-beforeEach(() => {
-  delete process.env.PRODUCTION;
-});
+import { createIngress } from "./createIngress";
 
 test("should create a dev ingress", () => {
   expect(
     createIngress({
       hosts: ["sample.dev2.fabrique.social.gouv.fr"],
+      isProduction: false,
       name: "my-ingress",
-      serviceName: "www",
-      servicePort: 80,
     })
   ).toMatchSnapshot();
 });
 
 test("should create a prod ingress", () => {
-  process.env.PRODUCTION = "true";
   expect(
     createIngress({
       hosts: ["sample.dev2.fabrique.social.gouv.fr"],
+      isProduction: true,
       name: "my-ingress",
       serviceName: "www",
-      servicePort: 80,
+      servicePortName: "http-foo",
     })
   ).toMatchSnapshot();
 });
@@ -34,9 +29,10 @@ test("should create an ingress with multiple hosts", () => {
         "sample.dev2.fabrique.social.gouv.fr",
         "www.sample.dev2.fabrique.social.gouv.fr",
       ],
+      isProduction: false,
       name: "my-ingress",
       serviceName: "www",
-      servicePort: 80,
+      servicePortName: "http-foo",
     })
   ).toMatchSnapshot();
 });